move resource_path to separate file
