+++ circleguard/utils.py
@@ -1,0 +1,16 @@
+from pathlib import Path
+import sys
+
+ROOT_PATH = Path(__file__).parent.absolute()
+
+def resource_path(str_path):
+    """
+    Returns a Path representing where to look for resource files for the program,
+    such as databases or images.
+
+    This location changes if the program is run from an application built with pyinstaller.
+    """
+
+    if hasattr(sys, '_MEIPASS'):  # being run from a pyinstall'd app
+        return Path(sys._MEIPASS) / Path(str_path)  # pylint: disable=no-member
+    return ROOT_PATH / Path(str_path)
